feat(#12): Add Dto and converter from Book class to BookDto.
